Conflicts:
	Holovibes/contrast_correction.cu
	Holovibes/contrast_correction.cuh

+++ Holovibes/hardware_limits.cc
@@ -1,5 +1,9 @@
 #include "hardware_limits.hh"
 
+#include <cuda_runtime.h>
+#include <device_launch_parameters.h>
+#include <cmath>
+
 unsigned int get_max_threads_1d()
 {
   cudaDeviceProp prop;
@@ -11,7 +15,7 @@ unsigned int get_max_threads_2d()
 {
   cudaDeviceProp prop;
   cudaGetDeviceProperties(&prop, 0);
-  return sqrt(prop.maxThreadsPerBlock);
+  return static_cast<unsigned int>(sqrt(prop.maxThreadsPerBlock));
 }
 
 unsigned int get_max_blocks()
